+++ src/main/java/com/dart/api/application/chat/ChatService.java
@@ -6,11 +6,11 @@ import org.springframework.transaction.annotation.Transactional;
 import com.dart.api.domain.auth.AuthUser;
 import com.dart.api.domain.chat.entity.ChatMessage;
 import com.dart.api.domain.chat.entity.ChatRoom;
-import com.dart.api.domain.chat.repo.ChatMessageRepository;
-import com.dart.api.domain.chat.repo.ChatRoomRepository;
+import com.dart.api.domain.chat.repository.ChatMessageRepository;
+import com.dart.api.domain.chat.repository.ChatRoomRepository;
 import com.dart.api.domain.gallery.entity.Gallery;
 import com.dart.api.domain.member.entity.Member;
-import com.dart.api.domain.member.repo.MemberRepository;
+import com.dart.api.domain.member.repository.MemberRepository;
 import com.dart.api.dto.chat.request.ChatMessageCreateDto;
 import com.dart.api.infrastructure.websocket.MemberSessionRegistry;
 import com.dart.global.error.exception.NotFoundException;
