+++ src/main/java/com/dart/api/application/gallery/GalleryService.java
@@ -20,10 +20,10 @@ import com.dart.api.domain.auth.AuthUser;
 import com.dart.api.domain.gallery.entity.Cost;
 import com.dart.api.domain.gallery.entity.Gallery;
 import com.dart.api.domain.gallery.entity.Hashtag;
-import com.dart.api.domain.gallery.repo.GalleryRepository;
-import com.dart.api.domain.gallery.repo.HashtagRepository;
+import com.dart.api.domain.gallery.repository.GalleryRepository;
+import com.dart.api.domain.gallery.repository.HashtagRepository;
 import com.dart.api.domain.member.entity.Member;
-import com.dart.api.domain.member.repo.MemberRepository;
+import com.dart.api.domain.member.repository.MemberRepository;
 import com.dart.api.dto.gallery.request.CreateGalleryDto;
 import com.dart.api.dto.gallery.request.DeleteGalleryDto;
 import com.dart.api.dto.gallery.response.GalleryAllResDto;
