+++ src/main/java/com/dart/api/domain/payment/repository/PaymentRepository.java
@@ -1,4 +1,4 @@
-package com.dart.api.domain.payment.repo;
+package com.dart.api.domain.payment.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
