+++ src/main/java/com/dart/api/domain/review/repository/ReviewRepository.java
@@ -1,4 +1,4 @@
-package com.dart.api.domain.review.repo;
+package com.dart.api.domain.review.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
